Split layer norm into its own class
